Move input functions into `input` submodule
